requirements.txt
image==1.5.33
numpy>=1.21
python-vlc>=3.0.0
Pillow>=10.0.0     
numpy>=1.21 
Kivy>=2.3.1
kivymd==1.2.0
dynaconf>=3.2.11